dto에 tag, content 필드 추가
