+++ src/main/java/org/sopt/dto/post/req/PostUpdateRequest.java
@@ -1,0 +1,7 @@
+package org.sopt.dto.post.req;
+
+public record PostUpdateRequest(
+        String title,
+        String content
+) {
+}
