--- src/main/java/org/sopt/dto/req/PostUpdateRequest.java
@@ -1,6 +1,0 @@
-package org.sopt.dto.req;
-
-public record PostUpdateRequest(
-        String title
-) {
-}
