--- src/main/java/org/sopt/dto/res/PostDetailResponse.java
@@ -1,10 +1,0 @@
-package org.sopt.dto.res;
-
-public record PostDetailResponse(
-        Long contentId,
-        String title
-) {
-    public static  PostDetailResponse of(final Long contentId, final String title) {
-        return new PostDetailResponse(contentId, title);
-    }
-}
