--- src/main/java/org/sopt/dto/res/PostUpdateResponse.java
@@ -1,10 +1,0 @@
-package org.sopt.dto.res;
-
-public record PostUpdateResponse(
-        Long contentId,
-        String title
-) {
-    public static  PostUpdateResponse of(final Long contentId, final String title) {
-        return new PostUpdateResponse(contentId, title);
-    }
-}
